Refactor tests and cleanup across modules
Parameterize renderer/datamodel tests, replace sleep with Awaitility, and move RequestUserUtil tests to Mockito; update related security configs and exception handling, tweak selenium helpers and analytics ordering, and add Awaitility/Mockito dependencies.

+++ gedbrowser/src/test/java/org/schoellerfamily/gedbrowser/selenium/base/PageWaiter.java
@@ -21,7 +21,7 @@ public interface PageWaiter {
      * @param driver the driver
      * @param multiplier timeout multiplier
      */
-    void waitForPageLoaded(WebDriver driver, int multiplier);
+    void waitForPageLoaded(WebDriver driver, long multiplier);
 
     /**
      * Wait for the page to load. Different implementations for different
